requirements.txt
livekit
sounddevice
numpy
scipy
# gtts
pyttsx3
openai-whisper
python-dotenv